fix: repair build after merging 41 PRs (missing imports, storage fns, types)
Several pre-existing and merge-introduced issues left the contract crate
non-compiling:

- lib.rs was missing its top-level `use soroban_sdk::{...}` import entirely
- types::Error was still the old 9-variant enum; unified it with the
  complete 47-variant enum in errors.rs via re-export
- StorageKey exceeded the #[contracttype] 50-variant hard limit (55
  variants); consolidated 5 rarely-combined singleton config keys
  (multisig TTL, rate limit config, decay config, council timelock delay,
  storage limits) into one `MiscConfig` entry to get back under the cap
- Added several Storage accessor functions that callers referenced but
  that were never defined (get/set_multisig_proposal, get/set_rate_limit_config,
  get/set_last_issuance_time, get/set_limits, get/add_endorsement(s),
  get_global_stats(_raw))
- Added missing types (IssuerStats, RevocationList, RevocationListFormat)
  and an Attestation.jurisdiction field that several modules already
  expected
- Added Attestation::generate_bridge_id and fixed a bundle.rs XDR-encoding
  import gap
- Fixed a handful of usize/u32 cast mismatches and Option/Result call-site
  bugs surfaced once the crate started compiling again
- Updated src/test.rs call sites to match actual client method names/
  signatures (soroban's generated client unwraps Result<T, Error> to T,
  panicking on error, rather than returning Result)

cargo build now succeeds with zero errors; cargo test --lib compiles and
407/429 tests pass. Regenerated test_snapshots/ to match.

+++ src/validation.rs
@@ -263,12 +263,12 @@ impl Validation {
             if let Some(meta) = metadata {
                 let len = meta.len();
                 if let Some(min) = constraints.min_metadata_len {
-                    if len < min as usize {
+                    if len < min {
                         return Err(Error::ConstraintViolation);
                     }
                 }
                 if let Some(max) = constraints.max_metadata_len {
-                    if len > max as usize {
+                    if len > max {
                         return Err(Error::ConstraintViolation);
                     }
                 }
